page pointers are not always 64bit

+++ src/core/memory.h
@@ -1,4 +1,4 @@
 int      lock_memory();
 void    *allocate_huge_page(int size);
-uint64_t phys_page(uint64_t virt_page);
+intptr_t phys_page(intptr_t virt_page);
